feat(login page): separate layout from login's page
