+++ lib/pages/login_user_page/commands/login_user_desktop_commands.dart
@@ -1,0 +1,5 @@
+import 'package:volume_vault/pages/login_user_page/commands/login_user_strategy.dart';
+
+class LoginUserDesktopCommands extends LoginUserStrategy {
+  //There is no specific implementation for desktop.
+}
