+++ lib/pages/login_user_page/commands/login_user_mobile_commands.dart
@@ -1,0 +1,5 @@
+import 'package:volume_vault/pages/login_user_page/commands/login_user_strategy.dart';
+
+class LoginUserMobileCommands extends LoginUserStrategy {
+  //There is no specific implementation for mobile.
+}
